+++ src/Tokenizer.cpp
@@ -14,12 +14,13 @@ TokenizerPtr Tokenizer::create(List spec) {
     char delim = as<char>(spec["delim"]);
     char quote = as<char>(spec["quote"]);
     std::vector<std::string> na = as<std::vector<std::string> >(spec["na"]);
+    std::string comment = as<std::string>(spec["comment"]);
     bool trimWs = as<bool>(spec["trim_ws"]);
     bool escapeDouble = as<bool>(spec["escape_double"]);
     bool escapeBackslash = as<bool>(spec["escape_backslash"]);
 
     return TokenizerPtr(new
-      TokenizerDelim(delim, quote, na, trimWs, escapeBackslash, escapeDouble)
+      TokenizerDelim(delim, quote, na, comment, trimWs, escapeBackslash, escapeDouble)
     );
   } else if (subclass == "tokenizer_fwf") {
     std::vector<int>